Super Wednesday part 2:
Lowered the accuracy on turnFast. We are spending too much time on turning and spitting out blues, which we are gonna need to fix.
Added an accuracy parameter to turnTo
Fixed the driveViaTimeGyroCamera method by reverting to what it used to be
Finished the auton, for real.

TODOS: Run over the pieces of code to make sure everything is perfect. We still need to cut about 20 seconds off, and fix the accuracy of things.